moving ChangeHistory from the model to the viewmodel, since it now provides change notification.
